backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pyjwt
passlib[bcrypt]
supabase
python-dotenv